tests: extract lifecycle and stack behavior modules
