+++ tests/test_lifecycle.h
@@ -1,0 +1,13 @@
+#ifndef TEST_LIFECYCLE_H
+#define TEST_LIFECYCLE_H
+
+void test_main_current(void);
+void test_add_unbound_phase1(void);
+void test_simple(void);
+void test_simple_create(void);
+void test_simple_create_and_run(void);
+void test_create_previous(void);
+void test_previous_cleared_on_manual_delete(void);
+void test_switch_new(void);
+
+#endif
