+++ tests/test_stack.h
@@ -1,0 +1,7 @@
+#ifndef TEST_STACK_H
+#define TEST_STACK_H
+
+void test_stack_further(void);
+void test_stack_far_isolation(void);
+
+#endif
