+++ tests/tests.c
@@ -6,9 +6,11 @@
 #include "tealet.h"
 #include "tealet_extras.h"
 #include "test_harness.h"
+#include "test_lifecycle.h"
 #include "test_lock_helpers.h"
 #include "test_locking.h"
 #include "test_resilience.h"
+#include "test_stack.h"
 #include "test_stress.h"
 #include "test_transfer.h"
 
@@ -248,7 +250,7 @@ tealet_t *tealet_new_native_call(tealet_t *m, tealet_run_t run, void **parg, voi
 }
 
 /* create a tealet or stub low on the stack */
-static int tealet_new_descend(tealet_t *t, tealet_t **out, int level, tealet_run_t run, void **parg, void *stack_far) {
+int tealet_new_descend(tealet_t *t, tealet_t **out, int level, tealet_run_t run, void **parg, void *stack_far) {
   int boo[10];
   boo[9] = 0;
   (void)boo;
@@ -347,7 +349,6 @@ static int stub_new3(tealet_t *t, tealet_t **out, tealet_run_t run, void **parg,
 
 typedef int (*t_new)(tealet_t *, tealet_t **, tealet_run_t, void **, void *);
 static t_new newarray[] = {tealet_new_rnd, stub_new, stub_new2, stub_new3};
-static tealet_t *(*tealet_new_native)(tealet_t *, tealet_run_t, void **, void *) = tealet_new_native_call;
 
 static t_new get_new() {
   if (newmode >= 0)
@@ -359,289 +360,6 @@ static t_new get_new() {
 
 int tealet_test_new_dispatch(tealet_t *t, tealet_t **out, tealet_run_t run, void **parg, void *stack_far) {
   return TEALET_TEST_NEW(t, out, run, parg, stack_far);
-}
-
-/************************************************************/
-
-void test_main_current(void) {
-  init_test();
-  fini_test();
-}
-
-static void test_stack_further_inner(void *outer_addr) {
-  int inner_local;
-  void *further = tealet_stack_further(outer_addr, &inner_local);
-  assert(further == outer_addr);
-}
-
-typedef struct stack_far_case_t {
-  int value;
-} stack_far_case_t;
-
-typedef struct stack_far_run_arg_t {
-  tealet_t *main;
-  stack_far_case_t *shared;
-  int before;
-  int after;
-} stack_far_run_arg_t;
-
-static tealet_t *test_stack_far_isolation_run(tealet_t *current, void *arg) {
-  stack_far_run_arg_t *run_arg = (stack_far_run_arg_t *)arg;
-  assert(current != run_arg->main);
-  assert(run_arg->shared->value == run_arg->before);
-  run_arg->shared->value = run_arg->after;
-  tealet_switch(run_arg->main, NULL, TEALET_XFER_DEFAULT);
-  assert(run_arg->shared->value == run_arg->after);
-  return run_arg->main;
-}
-
-static tealet_t *test_stack_far_isolation_parent(tealet_t *current, void *arg) {
-  stack_far_case_t shared;
-  stack_far_run_arg_t *run_arg;
-  tealet_t *child;
-  void *child_arg;
-  void *stack_far;
-  (void)arg;
-
-  shared.value = 11;
-  run_arg = (stack_far_run_arg_t *)tealet_malloc(current, sizeof(*run_arg));
-  assert(run_arg != NULL);
-  run_arg->main = current;
-  run_arg->shared = &shared;
-  run_arg->before = 11;
-  run_arg->after = 99;
-  child_arg = run_arg;
-
-  stack_far = tealet_stack_further(&shared, (void *)(&shared + 1));
-  child = tealet_new_native(current, test_stack_far_isolation_run, &child_arg, stack_far);
-  assert(child != NULL);
-
-  /* Child already ran once during creation (RUN_SWITCH): it wrote its private
-   * copy and switched back to parent.
-   */
-  assert(shared.value == 11);
-
-  /* Resume child: it confirms its private value, then returns/exits to parent. */
-  tealet_switch(child, NULL, TEALET_XFER_DEFAULT);
-
-  /* Child has exited; explicit delete is still required. */
-  tealet_delete(child);
-
-  /* Parent value remains unchanged throughout. */
-  assert(shared.value == 11);
-  tealet_free(current, run_arg);
-  return g_main;
-}
-
-void test_stack_further(void) {
-  int a_local;
-  int b_local;
-  int outer_local;
-  void *a;
-  void *b;
-  void *further_ab;
-  void *further_ba;
-  init_test();
-  a = &a_local;
-  b = &b_local;
-  further_ab = tealet_stack_further(a, b);
-  further_ba = tealet_stack_further(b, a);
-  assert(further_ab == a || further_ab == b);
-  assert(further_ba == a || further_ba == b);
-  assert(further_ab == further_ba);
-  assert(tealet_stack_further(a, a) == a);
-  assert(tealet_stack_further(b, b) == b);
-  assert(tealet_stack_further(further_ab, a) == further_ab);
-  assert(tealet_stack_further(further_ab, b) == further_ab);
-
-  /* Cross-frame check: caller frame local should be farther than callee local
-   */
-  test_stack_further_inner(&outer_local);
-  fini_test();
-}
-
-/* test that stack isolation works if tealets are created with an extended
- * stack. */
-void test_stack_far_isolation(void) {
-  tealet_t *parent;
-
-  init_test();
-  parent = tealet_new_native(g_main, test_stack_far_isolation_parent, NULL, NULL);
-  assert(parent != NULL);
-  tealet_delete(parent);
-  fini_test();
-}
-
-void test_add_unbound_phase1(void) {
-  tealet_t *unbound;
-  tealet_t *copy;
-  int rc;
-
-  init_test();
-  unbound = tealet_new(g_main);
-  assert(unbound != NULL);
-  assert(tealet_status(unbound) == TEALET_STATUS_NEW);
-  assert(tealet_get_far(unbound) == NULL);
-
-  copy = tealet_duplicate(unbound);
-  assert(copy != NULL);
-  assert(tealet_status(copy) == TEALET_STATUS_NEW);
-  assert(tealet_get_far(copy) == NULL);
-
-  rc = tealet_switch(unbound, NULL, TEALET_XFER_DEFAULT);
-  assert(rc == TEALET_ERR_INVAL);
-
-  tealet_delete(copy);
-  tealet_delete(unbound);
-  fini_test();
-}
-
-/************************************************************/
-
-tealet_t *test_simple_run(tealet_t *t1, void *arg) {
-  tealet_t *prev_current;
-  unsigned int origin_current;
-  unsigned int origin_main;
-  (void)arg;
-  assert(t1 != g_main);
-  origin_current = tealet_get_origin(t1);
-  origin_main = tealet_get_origin(g_main);
-  assert((origin_current & TEALET_ORIGIN_MAIN_LINEAGE) == 0);
-  assert((origin_current & TEALET_ORIGIN_FORK) == 0);
-  assert((origin_main & TEALET_ORIGIN_MAIN_LINEAGE) != 0);
-  assert((origin_main & TEALET_ORIGIN_FORK) == 0);
-  prev_current = tealet_previous(t1);
-  assert(prev_current == t1->main);
-  status = 1;
-  return g_main;
-}
-
-void test_simple(void) {
-  tealet_t *t;
-  init_test();
-  t = tealet_new_native_call(g_main, test_simple_run, NULL, NULL);
-  assert(t != NULL);
-  assert(status == 1);
-  tealet_delete(t);
-  fini_test();
-}
-
-void test_simple_create(void) {
-  tealet_t *t;
-  init_test();
-  t = NULL;
-  assert(tealet_spawn(g_main, &t, test_simple_run, NULL, NULL, TEALET_RUN_DEFAULT) == 0);
-  assert(status == 0);
-  tealet_delete(t);
-  fini_test();
-}
-
-void test_simple_create_and_run(void) {
-  tealet_t *t;
-  init_test();
-  t = NULL;
-  assert(tealet_spawn(g_main, &t, test_simple_run, NULL, NULL, TEALET_RUN_DEFAULT) == 0);
-  tealet_switch(t, NULL, TEALET_XFER_DEFAULT);
-  assert(status == 1);
-  assert(tealet_previous(g_main) == t);
-  tealet_delete(t);
-  assert(tealet_previous(g_main) == NULL);
-  fini_test();
-}
-
-/* Test that tealet_previous() is correct inside run function for tealet_create
- */
-static tealet_t *test_create_previous_run(tealet_t *t1, void *arg) {
-  /* When first switched to via tealet_switch(), previous should be main */
-  assert(tealet_previous(t1) == t1->main);
-  status = 42;
-  return g_main;
-}
-
-void test_create_previous(void) {
-  tealet_t *t;
-  init_test();
-  /* Create tealet without running it */
-  t = NULL;
-  assert(tealet_spawn(g_main, &t, test_create_previous_run, NULL, NULL, TEALET_RUN_DEFAULT) == 0);
-  assert(status == 0);
-  /* Now switch to it - it should see main as previous */
-  tealet_switch(t, NULL, TEALET_XFER_DEFAULT);
-  assert(status == 42); /* Verify it ran */
-  assert(tealet_previous(g_main) == t);
-  tealet_delete(t);
-  assert(tealet_previous(g_main) == NULL);
-  fini_test();
-}
-
-static tealet_t *test_previous_manual_delete_run(tealet_t *t1, void *arg) {
-  (void)arg;
-  tealet_switch(t1->main, NULL, TEALET_XFER_DEFAULT);
-  return t1->main;
-}
-
-void test_previous_cleared_on_manual_delete(void) {
-  tealet_t *t;
-
-  init_test();
-  t = NULL;
-  assert(tealet_spawn(g_main, &t, test_previous_manual_delete_run, NULL, NULL, TEALET_RUN_DEFAULT) == 0);
-  assert(t != NULL);
-
-  tealet_switch(t, NULL, TEALET_XFER_DEFAULT);
-  assert(tealet_previous(g_main) == t);
-
-  tealet_delete(t);
-  assert(tealet_previous(g_main) == NULL);
-  fini_test();
-}
-
-/* 1 is high on the stack.  We then create 2 lower on the stack */
-/* the execution is : m 1 m 2 1 m 2 m */
-tealet_t *test_switch_new_1(tealet_t *t1, void *arg) {
-  tealet_t *caller = (tealet_t *)arg;
-  tealet_t *stub;
-  int result;
-  /* switch back to the creator */
-  tealet_switch(caller, NULL, TEALET_XFER_DEFAULT);
-  /* now we want to trample the stack */
-  stub = NULL;
-  result = tealet_new_descend(t1, &stub, 50, NULL, NULL, NULL);
-  assert(result == 0);
-  assert(stub != NULL);
-  tealet_delete(stub);
-  /* and back to main */
-  return g_main;
-}
-
-tealet_t *test_switch_new_2(tealet_t *t2, void *arg) {
-  tealet_t *target = (tealet_t *)arg;
-  /* switch to tealet 1 to trample the stack*/
-  target->extra = (void *)t2;
-  tealet_switch(target, NULL, TEALET_XFER_DEFAULT);
-
-  /* and then return to main */
-  return g_main;
-}
-
-void test_switch_new(void) {
-  tealet_t *tealet1, *tealet2;
-  void *arg;
-  int result;
-  init_test();
-  arg = (void *)tealet_current(g_main);
-  tealet1 = tealet_new_native_call(g_main, test_switch_new_1, &arg, NULL);
-  /* the tealet is now running */
-  arg = (void *)tealet1;
-  tealet2 = NULL;
-  result = tealet_new_descend(g_main, &tealet2, 4, test_switch_new_2, &arg, NULL);
-  assert(result == 0);
-  assert(tealet2 != NULL);
-  assert(tealet_status(tealet2) == TEALET_STATUS_ACTIVE);
-  tealet_switch(tealet2, NULL, TEALET_XFER_DEFAULT);
-  tealet_delete(tealet1);
-  tealet_delete(tealet2);
-  fini_test();
 }
 
 typedef struct extradata {
